+++ modules/apps/collaboration/document-library/document-library-test/src/testIntegration/java/com/liferay/document/library/verify/test/DLServiceVerifyProcessTest.java
@@ -16,7 +16,6 @@ package com.liferay.document.library.verify.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.asset.kernel.service.AssetEntryLocalServiceUtil;
-import com.liferay.document.library.kernel.exception.NoSuchFileVersionException;
 import com.liferay.document.library.kernel.model.DLFileEntry;
 import com.liferay.document.library.kernel.model.DLFileEntryMetadata;
 import com.liferay.document.library.kernel.model.DLFileEntryType;
@@ -28,10 +27,8 @@ import com.liferay.document.library.kernel.service.DLAppServiceUtil;
 import com.liferay.document.library.kernel.service.DLFileEntryLocalServiceUtil;
 import com.liferay.document.library.kernel.service.DLFileEntryMetadataLocalServiceUtil;
 import com.liferay.document.library.kernel.service.DLFileEntryTypeLocalServiceUtil;
-import com.liferay.document.library.kernel.service.DLFileVersionLocalServiceUtil;
 import com.liferay.document.library.kernel.service.DLFolderLocalServiceUtil;
 import com.liferay.document.library.kernel.service.DLTrashServiceUtil;
-import com.liferay.document.library.kernel.store.DLStoreUtil;
 import com.liferay.document.library.kernel.util.DLValidator;
 import com.liferay.dynamic.data.mapping.io.DDMFormXSDDeserializer;
 import com.liferay.dynamic.data.mapping.kernel.DDMForm;
@@ -61,7 +58,6 @@ import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.FileUtil;
 import com.liferay.portal.kernel.util.HashMapDictionary;
 import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.test.randomizerbumpers.TikaSafeRandomizerBumper;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
@@ -259,44 +255,6 @@ public class DLServiceVerifyProcessTest extends BaseVerifyProcessTestCase {
 		Assert.assertNotNull(
 			AssetEntryLocalServiceUtil.fetchEntry(
 				DLFileEntry.class.getName(), fileEntryId));
-	}
-
-	@Test
-	public void testDLFileEntryWithNoDLFileVersionAndNoFile() throws Exception {
-		DLFileEntry dlFileEntry = addDLFileEntry();
-
-		DLFileVersion dlFileVersion = dlFileEntry.getLatestFileVersion(true);
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_group, TestPropsValues.getUserId());
-
-		DLFileEntryLocalServiceUtil.updateStatus(
-			TestPropsValues.getUserId(), dlFileVersion.getFileVersionId(),
-			WorkflowConstants.STATUS_APPROVED, serviceContext, new HashMap<>());
-
-		DLFileVersionLocalServiceUtil.deleteDLFileVersion(
-			dlFileEntry.getFileVersion());
-
-		DLStoreUtil.deleteFile(
-			dlFileEntry.getCompanyId(), dlFileEntry.getDataRepositoryId(),
-			dlFileEntry.getName(), dlFileEntry.getVersion());
-
-		try {
-			dlFileEntry.getFileVersion();
-
-			Assert.fail();
-		}
-		catch (NoSuchFileVersionException nsfve) {
-		}
-
-		try (ConfigurationTemporarySwapper configurationTemporarySwapper =
-				_getConfigurationTemporarySwapper("fileExtensions", ".jpg")) {
-
-			doVerify();
-
-			Assert.assertNotNull(dlFileEntry.getFileVersion());
-		}
 	}
 
 	@Test