+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/form_builder_field_support.js
@@ -84,7 +84,20 @@ AUI.add(
 						}
 
 						if (fieldName === 'name') {
-							settingsFormFieldContext.value = copy.generateFieldName(instance.get('fieldName'));
+							var oldFieldName = instance.get('fieldName');
+
+							var newFieldName = Lang.sub(
+								Liferay.Language.get('copy-of-x'),
+								[
+									oldFieldName
+								]
+							);
+
+							newFieldName = Lang.String.camelize(newFieldName, ' ');
+
+							newFieldName = newFieldName.replace(/\s/g, '');
+
+							settingsFormFieldContext.value = copy.generateFieldName(newFieldName);
 						}
 					}
 				);