+++ modules/apps/web-experience/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/FragmentsEditor.es.js
@@ -43,11 +43,17 @@ class FragmentsEditor extends Component {
 			[...this.sidebarTabs];
 
 		this.on('languageIdChanged', this._handleLanguageIdChange);
+	}
 
-		this._translationStatus = this._getTranslationStatus(
-			Object.keys(this.availableLanguages).filter(languageId => languageId !== '_INJECTED_DATA_'),
-			this._getEditableValues(this.fragmentEntryLinks)
-		);
+	rendered(firstRender) {
+		if (firstRender) {
+			this._translationStatus = this._getTranslationStatus(
+				Object
+					.keys(this.availableLanguages)
+					.filter(languageId => languageId !== '_INJECTED_DATA_'),
+				this._getEditableValues()
+			);
+		}
 	}
 
 	/**
@@ -149,12 +155,11 @@ class FragmentsEditor extends Component {
 	 */
 
 	_getEditableValues() {
-		return Object
-			.keys(this.fragmentEntryLinks)
+		return this.fragmentEntryLinks
 			.map(
-				fragmentEntryLinkId => {
+				fragmentEntryLink => {
 					const component = this._getFragmentEntryLinkComponent(
-						fragmentEntryLinkId
+						fragmentEntryLink.fragmentEntryLinkId
 					);
 
 					return component ? component.getEditableValues() : null;
@@ -468,6 +473,13 @@ class FragmentsEditor extends Component {
 				this._sidebarTabs = [...this.sidebarTabs];
 			}
 
+			this._translationStatus = this._getTranslationStatus(
+				Object
+					.keys(this.availableLanguages)
+					.filter(languageId => languageId !== '_INJECTED_DATA_'),
+				this._getEditableValues()
+			);
+
 			this._deleteFragmentEntryLink(data.fragmentEntryLinkId);
 		}
 	}
@@ -609,7 +621,7 @@ class FragmentsEditor extends Component {
 
 					this._translationStatus = this._getTranslationStatus(
 						Object.keys(this.availableLanguages).filter(languageId => languageId !== '_INJECTED_DATA_'),
-						this._getEditableValues(this.fragmentEntryLinks)
+						this._getEditableValues()
 					);
 
 					const fragmentEntryLinkComponent = this.refs[`fragmentEntryLink_${fragmentEntryLink.fragmentEntryLinkId}`];